+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.4.0] - 2025-04-10
+
 ### Added
 
 - Allow setting the `traverse` state of an `entity.component_tags[type][entity](traverse=...)` attribute.